Stop unmeasured page sizes producing Infinity page numbers
Auto rows-per-page resolves to 0 until the fit is measured, and four sites
divided by it: the transmit history counter rendered 1 / Infinity on first
paint, and three timeline scrubs called setCurrentPage(Infinity). pageCount
and pageForOffset keep that guard in one place. transmit_history::query takes
u32 now because SQLite reads a negative LIMIT as no limit, so a stray sentinel
would have quietly returned the whole table.

+++ src-tauri/src/transmit_history.rs
@@ -170,7 +170,10 @@ pub fn count() -> i64 {
 }
 
 /// Return up to `limit` rows ordered by newest first, starting at `offset`.
-pub fn query(offset: i64, limit: i64) -> Vec<TransmitHistoryRow> {
+///
+/// Unsigned deliberately: SQLite reads a negative `LIMIT` as *no limit*, so a bad
+/// value would quietly return the whole table. `u32` makes serde reject it instead.
+pub fn query(offset: u32, limit: u32) -> Vec<TransmitHistoryRow> {
     let db = match DB.lock() {
         Ok(g) => g,
         Err(_) => return vec![],
@@ -264,7 +267,7 @@ pub fn find_offset(timestamp_us: i64) -> i64 {
 // ============================================================================
 
 #[tauri::command]
-pub fn transmit_history_query(offset: i64, limit: i64) -> Result<Vec<TransmitHistoryRow>, String> {
+pub fn transmit_history_query(offset: u32, limit: u32) -> Result<Vec<TransmitHistoryRow>, String> {
     Ok(query(offset, limit))
 }
